TCCC-197
git-svn-id: https://coder.topcoder.com/tcs/clients/cronos/components@123498 fb370eea-3af6-4597-97f7-f7400a59c12a

+++ studio_service/trunk/src/java/main/com/topcoder/service/studio/ContestData.java
@@ -89,7 +89,7 @@ public class ContestData implements Serializable {
      * Represents the duration In Hours
      * </p>
      */
-    private int durationInHours = -1;
+    private double durationInHours = -1;
 
     /**
      * <p>
@@ -371,7 +371,7 @@ public class ContestData implements Serializable {
      *
      * @return the durationInHours
      */
-    public int getDurationInHours() {
+    public double getDurationInHours() {
         return durationInHours;
     }
 
@@ -383,7 +383,7 @@ public class ContestData implements Serializable {
      * @param durationInHours
      *            the durationInHours to set
      */
-    public void setDurationInHours(int durationInHours) {
+    public void setDurationInHours(double durationInHours) {
         this.durationInHours = durationInHours;
     }
