+++ studio_service/trunk/src/java/main/com/topcoder/service/studio/ejb/StudioServiceBean.java
@@ -979,7 +979,7 @@ public class StudioServiceBean implements StudioService {
         Date startDate = getDate(data.getLaunchDateAndTime());
         result.setStartDate(startDate);
         if (startDate != null) {
-            result.setEndDate(new Date(startDate.getTime() + 60l * 60 * data.getDurationInHours()));
+            result.setEndDate(new Date((long)(startDate.getTime() + 60l * 60 * data.getDurationInHours())));
         }
 
         addContestConfig(result, contestPropertyShortSummaryId, data.getShortSummary());
